+++ src/tollfree_verifications.rs
@@ -422,8 +422,8 @@ impl<'a> TollfreeVerificationFields<'a> {
         validate_optional_enum("VettingProvider", self.vetting_provider)
     }
 
-    fn sensitive_values(self, creds: TwilioCreds<'a>, sid: Option<&'a str>) -> Vec<&'a str> {
-        let mut values = vec![creds.account_sid, creds.auth_token];
+    fn sensitive_values(self, creds: &'a TwilioCreds, sid: Option<&'a str>) -> Vec<&'a str> {
+        let mut values = vec![creds.account_sid(), creds.auth_token()];
         push_sensitive(&mut values, sid);
         push_sensitive(&mut values, self.business_name);
         push_sensitive(&mut values, self.business_website);
@@ -846,7 +846,7 @@ impl<'a> CreateTollfreeVerificationRequest<'a> {
         self.fields.form_params()
     }
 
-    fn sensitive_values(self, creds: TwilioCreds<'a>) -> Vec<&'a str> {
+    fn sensitive_values(self, creds: &'a TwilioCreds) -> Vec<&'a str> {
         self.fields.sensitive_values(creds, None)
     }
 }
@@ -884,7 +884,7 @@ impl<'a> UpdateTollfreeVerificationRequest<'a> {
         self.fields.form_params()
     }
 
-    fn sensitive_values(self, creds: TwilioCreds<'a>, sid: &'a str) -> Vec<&'a str> {
+    fn sensitive_values(self, creds: &'a TwilioCreds, sid: &'a str) -> Vec<&'a str> {
         self.fields.sensitive_values(creds, Some(sid))
     }
 }
@@ -990,8 +990,8 @@ impl<'a> ListTollfreeVerificationsRequest<'a> {
         }
     }
 
-    fn sensitive_values(self, creds: TwilioCreds<'a>) -> Vec<&'a str> {
-        let mut values = vec![creds.account_sid, creds.auth_token];
+    fn sensitive_values(self, creds: &'a TwilioCreds) -> Vec<&'a str> {
+        let mut values = vec![creds.account_sid(), creds.auth_token()];
         push_sensitive(&mut values, self.tollfree_phone_number_sid);
         push_sensitive(&mut values, self.external_reference_id);
         values.extend(self.trust_product_sids.iter().copied());
@@ -1466,8 +1466,8 @@ impl<'a> TollfreeVerificationsResource<'a> {
     ) -> Result<TwilioTollfreeVerificationPage, TwilioError> {
         async move {
             let sensitive_values = vec![
-                self.account.creds.account_sid,
-                self.account.creds.auth_token,
+                self.account.creds.account_sid(),
+                self.account.creds.auth_token(),
                 next_page_url,
             ];
             let resource = V1PageResource::TollfreeVerifications;
@@ -1657,8 +1657,8 @@ impl<'a> TollfreeVerificationResource<'a> {
 
     fn sensitive_values(self) -> Vec<&'a str> {
         vec![
-            self.account.creds.account_sid,
-            self.account.creds.auth_token,
+            self.account.creds.account_sid(),
+            self.account.creds.auth_token(),
             self.sid,
         ]
     }
@@ -1758,8 +1758,8 @@ impl<'a> BlockingTollfreeVerificationsResource<'a> {
         )
         .in_scope(|| {
             let sensitive_values = vec![
-                self.account.creds.account_sid,
-                self.account.creds.auth_token,
+                self.account.creds.account_sid(),
+                self.account.creds.auth_token(),
                 next_page_url,
             ];
             let resource = V1PageResource::TollfreeVerifications;
@@ -1934,8 +1934,8 @@ impl<'a> BlockingTollfreeVerificationResource<'a> {
 
     fn sensitive_values(self) -> Vec<&'a str> {
         vec![
-            self.account.creds.account_sid,
-            self.account.creds.auth_token,
+            self.account.creds.account_sid(),
+            self.account.creds.auth_token(),
             self.sid,
         ]
     }